add GTA dataset

+++ LeReS/Train/data/gta_dataset.py
@@ -119,7 +119,7 @@ class GTADataset(Dataset):
         joints_3d_world = self.info_npz['joints_3d_world'][anno_index]
         world2cam_trans = self.info_npz['world2cam_trans'][anno_index]
         intrinsics = self.info_npz['intrinsics'][anno_index]
-        focal_length = intrinsics[0][0]
+        focal_length = (intrinsics[0][0]).astype(np.float32)
         depth, invalid_depth, sem_mask = self.load_training_data(anno_index)
 
         rgb_aug = self.rgb_aug(rgb)
